requirements.txt
aerich==0.7.1
aiofiles==22.1.0
aiogram==3.0.0b5
aiohttp==3.8.3
aiosignal==1.3.1
aiosqlite==0.17.0
anyio==3.6.2
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.2.0
Babel==2.11.0
Brotli==1.0.9
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.4
Deprecated==1.2.13
dictdiffer==0.9.0
fastapi==0.89.1
frozenlist==1.3.3
h11==0.14.0
idna==3.4
iso8601==1.1.0
magic-filter==1.0.9
multidict==6.0.4
packaging==23.0
pycparser==2.21
pydantic==1.10.4
pyOpenSSL==22.1.0
pypika-tortoise==0.1.6
python-dotenv==0.21.0
pytz==2022.7.1
redis==4.3.4
sniffio==1.3.0
starlette==0.22.0
tomlkit==0.11.6
tortoise-orm==0.19.0
typing_extensions==4.4.0
uvicorn==0.20.0
wrapt==1.14.1
yarl==1.8.2
